added browser functionality in main login command with --no-browser capabilites backwards compatibility
